requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128/
torch==2.8.0+cu128
torchvision==0.23.0+cu128
numpy==2.4.3
Pillow==12.1.1
PyYAML==6.0.3
torch_optimizer==0.3.0
tqdm==4.67.1
requests==2.32.5
tensorboard>=2.12.3
scikit-learn==1.7.1
